requirements.txt
Django>3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
pandas>=1.4,<1.5
numpy==1.18.4
scikit-learn==0.24.2
joblib==0.16.0
markdown==3.4.1